requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.4.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.13
datasets==1.18.3
dill==0.4.0
einops==0.4.0
rouge==1.0.1
sentence-transformers==2.2.2
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
hf-xet==1.2.0
higher==0.2.1
huggingface-hub==0.36.0
hydra-core==1.2.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
MarkupSafe==2.1.5
matplotlib==3.6.1
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
networkx==3.3
nltk==3.7
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==25.0
pandas==2.3.3
pathlib_abc==0.1.1
pathy==0.11.0
pillow==11.3.0
preshed==3.0.12
propcache==0.4.1
pyarrow==10.0.1
pydantic==1.9.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.0.2
scipy==1.9.2
six==1.17.0
smart-open==6.4.0
spacy==3.4.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sympy==1.13.1
thinc==8.1.12
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.5.1+cu121 # --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.1
triton==3.1.0
typer==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==0.10.1
xxhash==3.6.0
yarl==1.22.0
